Split out temporary code style excludes
Splitting them out allows for plugins to use _test/phpcs.xml as the
basis for their own linting without having an overly permissive coding
standard.
Also, this makes it more obvious and painful that these are just
intended as temporary exceptions and should be actually fixed.

The rule `Generic.ControlStructures.InlineControlStructure.NotAllowed`
has been moved to the section with the temporary excludes and had its
comment removed as it is in fact in PSR-2 (and -12) and can also be
auto-fixed by phpcbf. See squizlabs/PHP_CodeSniffer#2744

The rule `PSR1.Classes.ClassDeclaration.MissingNamespace` has to remain
in the DokuWiki coding standard as the plugin base classes can currently
not reasonably be in namespaces.